Import webpki's SPKI testing code and test data as NPOTB.
Imported from webpki af55a9861eeac6be8bd4c062823b86f2534027da. This
will form the basis of the implementation and testing of SPKI parsing.
All of this code is existing code from webpki. The importing was done
by Marshall Pierce and was split out of another commit of his.