RSA/bigint: Clarify citation of "Manger's Attack revisited" paper.
Citing the paper 3x is overkill. Even citing it once is probably overkill
considering we've been designing everything under the assumption that fixed
precision constant-time arithmetic should be assumed to be required unless
documented.

At the same time, clarify which part of the paper we're actually citing,
since I had to look it up.